Organize Actual mode core module
