+++ libs/@hashintel/petrinaut-core/src/actual-mode/constants.ts
@@ -1,0 +1,2 @@
+export const ACTUAL_MODE_TIMELINE_TICK_MS = 500;
+export const ACTUAL_MODE_RECORDING_VERSION = 1;
